structure updates continued
