+++ src/configuration/mod.rs
@@ -1,1 +1,3 @@
 pub mod configuration;
+pub mod keybindings;
+pub mod styles;
